requirements.txt
aiohttp
fastapi
mistralai
openai
google-genai
pydantic
pydantic_settings
python-dotenv
uvicorn
beanie
motor
pymongo
